s3vectors-app/requirements.txt
Flask==2.3.2
boto3
Pillow
torch
timm
scikit-learn
sentence-transformers
numpy
